Move `into_json` and `log_response` to crate root
These will both be used by the async client. Move them out of sync
client in preparation.

Edit the logging function so it doesn't panic.
